refactor: make code smaller in a modularize pattern
